rasa_requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==23.2.1
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
bidict==0.23.1
boto3==1.34.51
botocore==1.34.51
CacheControl==0.12.14
cachetools==4.2.4
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==42.0.5
cycler==0.12.1
dask==2021.11.2
dnspython==1.16.0
docopt==0.6.2
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==2.0.7
fsspec==2024.2.0
future==1.0.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.5.3
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
locket==1.0.0
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.3.4
mattermostwrapper==2.2
msgpack==1.0.7
multidict==5.2.0
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
partd==1.4.1
pillow==10.2.0
prompt-toolkit==2.0.10
protobuf==3.19.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.1.1
pyrsistent==0.20.0
pyTelegramBotAPI==3.8.3
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2021.3
PyYAML==6.0.1
questionary==1.10.0
randomname==0.1.5
rasa==3.1.0
rasa-sdk==3.1.3
redis==3.5.3
regex==2021.8.28
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rocketchat-API==1.16.0
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.10.1
sentry-sdk==1.3.1
serpapi==0.1.5
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.51
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow==2.7.4
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-text==2.7.3
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.3.0
toolz==0.12.1
tqdm==4.66.2
twilio==6.50.1
typeguard==2.13.3
typing-extensions==3.10.0.2
typing-utils==0.1.0
tzlocal==2.1
ujson==4.3.0
urllib3==1.26.18
uvloop==0.19.0
wcwidth==0.2.13
webexteamssdk==1.6.1
websockets==10.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
